refactor: correção na tentativa de abertura de arquivos que não podem ser abertos mais de uma vez.
