+++ src/components/ui/index.ts
@@ -1,0 +1,6 @@
+import Button from './Button';
+import Slider from './Slider';
+
+export { Button, Slider };
+
+export default { Button, Slider };
