+++ src/store/slices/window/windowSlice.ts
@@ -63,10 +63,16 @@ const windowSlice = createSlice({
         );
         if (
           foundSameWindow &&
+          foundSameWindow.position &&
+          foundSameWindow.size &&
           foundSameWindow.windowId !== undefined &&
           foundSameWindow.windowState
         ) {
           state.focusedWindow = foundSameWindow.windowId;
+          foundSameWindow.position.lastX = foundSameWindow.position.x;
+          foundSameWindow.position.lastY = foundSameWindow.position.y;
+          foundSameWindow.size.lastWidth = foundSameWindow.size.width;
+          foundSameWindow.size.lastHeight = foundSameWindow.size.height;
           foundSameWindow.windowState.requests.restore = true;
           return;
         }
